Add test dependencies on Humble branch. (#516)
ament_lint_auto nor ament_lint_common are not listed as dependencies for
smacc2_msgs. ament_lint_auto is required due to the `find_package` call
and while `ament_lint_common` is not mentioned by name, ament_lint_auto
will only work if there are ament_lint packages present for it to
configure.

This is causing the build failures mentioned in
https://github.com/robosoft-ai/SMACC2/issues/508#issuecomment-1591685849 robosoft-ai/SMACC2@d198fc6673c89f1e75a497a3f41bb1c76251c9b1